fix: handle nested directories as preprocessing input
Example: input directory contains files `project.json` and `a/project.json`.
In this case the output would only be `project.txt` since one of the two results
would overwrite the other one.
This is fixed by recreating the directory structure of the input directory also
in the output one. I.e. the output directory would now contain `project.txt` and
`a/project.txt`.