refactor(tokenizer): use infix order for binary operators
The operators
`+`, `-`, `*`, `/`, `mod`, `>`, `<`, `==`, `and`, `or`
were tokenized like `operator_add (operand1) (operand2)` before.
This commit changes the order to `(operand1) operator_add (operand2)`.

In Scratch the operand order of blocks containing text depends on the
user interface language and the corresponding block text translation.
Therefore, using the Polish notation for those blocks at least
achieves a consistent tokenisation and acts similar to function calls
as seen in other languages.

However, for operators the order is always the same regardless of
natural language of the user interface. Therefore, I propose to keep
the infix notation for those operators. This order then also mimics
the one as used in most programming languages (except for Lisp-like
ones?). Especially when using the Scratch token sequences to fine-tune
a larger model that has been pre-trained on regular code this should
be beneficial.